+++ packages/frontend-logic/src/shared/reducers/fixedLocsReducer.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { handleAxiosError } from '../utils/errorHandler';
+import { handleAxiosError } from '../../utils/errorHandler';
 import { AppDispatch } from '../store';
-import fixedLocRouter from '../services/fixedLoc';
+import fixedLocService from '../services/fixedLoc';
 import { ApplicationError, FixedLocDT, isFixedLocDT, SafeyAny } from '@m-cafe-app/utils';
 import { TFunction } from '../hooks/useTranslation';
 
@@ -34,7 +34,7 @@ export const { setFixedLocs } = fixedLocSlice.actions;
 export const initFixedLocs = (t: TFunction) => {
   return async (dispatch: AppDispatch) => {
     try {
-      const fixedLocs = await fixedLocRouter.getLocs();
+      const fixedLocs = await fixedLocService.getLocs();
       if (!Array.isArray(fixedLocs)) throw new ApplicationError('Server has sent wrong data', { current: fixedLocs });
       for (const fixedLoc of fixedLocs) {
         if (!isFixedLocDT(fixedLoc)) throw new ApplicationError('Server has sent wrong data', { all: fixedLocs, current: fixedLoc as SafeyAny });
